backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
Werkzeug==3.0.1
SQLAlchemy==2.0.46
python-dotenv==1.0.0
scikit-learn
pandas==2.2.3
nltk==3.8.1
cloudpickle==2.2.1
google-generativeai==0.3.0